+++ CHANGELOG.md
@@ -11,8 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Waterfall: Optimizing the size of the comment text
+- Risk-review: Legend moved from the tooltip to the main visualization
 
 ### Fixed
+- Problem with select list overlapped by other panels
+- Risk-review: Fixed processing of large numbers
 
 ## [1.144.0]